backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
reportlab==4.0.4
django-filter==23.2
gunicorn==20.1.0
python-dotenv==1.0.0
pillow==10.0.0
djangorestframework==3.14.0
djoser==2.2.0
drf-extra-fields==3.7.0